refactor: rename module `two_fa` to `two_factor_auth`
